refactor: WaitingValidator 위치 수정
