+++ src/test/java/roomescape/reservation/presentation/ReservationControllerTest.java
@@ -34,9 +34,9 @@ import roomescape.global.auth.AdminInterceptor;
 import roomescape.global.exception.ReservationErrorCode;
 import roomescape.global.exception.customException.BusinessException;
 import roomescape.global.exception.customException.EntityNotFoundException;
+import roomescape.reservation.application.ReservationService;
 import roomescape.reservation.application.dto.ReservationCreateCommand;
 import roomescape.reservation.application.dto.ReservationUpdateCommand;
-import roomescape.reservation.application.ReservationService;
 import roomescape.reservation.application.dto.UserReservationResult;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservationTime.domain.ReservationTime;
